First, refactor CollectdITest to make "getServerData" logic re-usable
